refactor(TagExplorer): Extract ExploreHeader component
